Add unit tests for cubic splines
Solve for the coefficients using a straightforward (but inefficient) derivation
from the defining equations.

The functions in eqn_manip.py perform the equation manipulations one might use
when working by hand to put equations in a particular form - eg, moving all of
one type of term to one side, in order to set up a matrix equation.

Test the coefficient solver and the evaluation class (OneDimCubicSpline).