LCAO and PyscfToQmcpack typos and inconsistencies
PyscfToQmcpack was working with a variable called SpinUnRestricted,
which is not ever read by LCAOH5Parser. So it would always default to
spin restricted calculations since that is the default by LCAO parser.
This was causing problems for the crystal converter when wanting
unrestricted calculations.

Also, there were typos in the variable names being read from the HDF5
for SpinRestricted