Inconsistent use of spin multiplicity.
The LCAOH5Parser, which is based on QMCGaussianParserBase, uses
SpinMultiplicty, which is 2S+1. It seems that the PyscfToQmcpack uses
the spin variable from the cell object, which is only 2S.

The solution here keeps PyscfToQmcpack writing the "spin", which is 2S.
Then LCAOH5Parser will read that as ds, and set SpinMultiplicity to
2S+1, which is necessary since that is what QMCGaussianParserBase
assumes. Another easy solution is to just make sure the hdf5 files store
spin_multiplity instead and LCAOH5Parser will just read that directly.
The crystal converter needs this resolved, so I'm trying to get this
cleared up.