Remove index from one dim spline state
The Loc variable carried state about the index location on the one dimensional
grid.   Change the 'locate' function to return the index instead.

This might result in duplicate calls to 'locate' in a few places, but that should
not affect performance (much?).  Those duplicate calls will be removed in the
near future as they index information will be combined with calls to updateSecondOrder
and updateForQunitic.