Remove state from quintic spline evalution
Add a function to get both the index and the distance from the grid point.
This is the only computation performed by updateForQuintic, so removing
that function should not have a performance impact.

Also, this code is only used for evaluation in the AoS code - the SoA version
is in QMCWavefunctions/lcao/MultiQuinticSpline1D.h, and is unaffected by this change.

The quinticInterpolate series of functions were moved to OneDimQuinticSpline from
OneDimGridBase.  They were also renamed to have separate names (based on the
highest derivative computed), rather than overloading a single name.