squeue correction for nexus
This PR fixes a small issue with squeue usage on the NERSC machines.
Evidently, the NERSC machines used to print squeue information in a
different order, but do so no longer. This fixes the reading of the
squeue output to gather the correct information. In the case that the
squeue REASON is longer than a single string (as is often the case on the NERSC
machines), now we only read the first 8 variables since the reason string is
never used anyway.

Additionally, I added the remaining squeue job states, because the NERSC
machines often have jobs in this various job states causing nexus to
fail.