Open HDF orbital file in read-only mode in PW code
If the wavefunction unit tests are run under MPI, they halt here opening the
orbital HDF file.  Change the mode to read-only so multiple nodes could open the
file simultaneously.
This code is not performance critical nor used much, so updating it to read the
HDF file properly in parallel is not a priority.