Expand minimum bracket search for cusp correction
This addresses a case where the starting function value is around 1e18,
and is very sharply peaked near the starting value. The step size is not
large enough to reach the minimum in 1000 iterations.
The algorithm succeeds if another step size increment is added at 1000 iterations
and the max iteration count raised to 10000.