Catch bracket errors during cusp minimization
The code to bracket a minimum throws an exception, which is not well
reported, espcially under MPI.  This catches the exception and aborts
the execution with a more visible message.

The effective local energy divides by the value of wavefunction correction.
If it is zero, the result is a NaN, which makes bracketing and minimization
run poorly.  This is a simple hack to prevent the denominator from being zero.