Minor performance optimization for cusp correction
Finding the optimal cusp correction parameters is a nested loop - finding phi0 in the
inner loop and Rc in the outer loop.  The original algorithm would always start phi0
in the inner loop from the value of phi at 0.0 for each new value of Rc.
This change saves the optimal value of phi0 from the previous Rc and uses it as the
starting value in the inner loop for the next Rc.  Assuming that the optimal phi0 doesn't
vary much with Rc, this should speed up the search in subsequent inner loops.