Replace distance table N vector with variables
The distance table uses the N vector for sizes of the source,
target, and walkers.  Replace this with individual variables.

N[SourceIndex] -> N_sources
N[VisitorIndex] -> N_targets
N[PairIndex] -> not used
N[WalkerIndex] -> N_walkers

The Ntargets variable was moved from the SoA distance table back
to the base distance table.  It was renamed to N_targets to be
slightly more compliant with the naming convention.