Update distance table test for SoA
Make the distance table test work with SoA code in addition to the
existing AoS code.  Also change the loop bounds to use the distance table
values, which should be the same as the particle object values (check added).

Update the DistanceTableData documentation for Distances and Displacements,
which has the source and target indices reversed.