Fix band read for hybrids
In this PR, a fix is made that avoids band read failure when a hybrid
functional is used in the calculation. During a hybrid calculation,
band data is printed multiple times in the output (once for each scf
convergence). The changes here make sure that nfound, index, and the
band objects are re-initialized each time '- SPIN UP -' is seen in
the output.