Fix timer crash in optimizer
Addresses #1854

The timers in QMCLinearOptimize are used in derived classes
QMCFixedSampleLinearOptimize and QMCCorrelatedSamplingLinearOptimize.
The pointers to the myTimers caused seg faults as the array is now
longer initialized.  Update the timers in those classes.

Also remove the myTimers array from the QMCDriver base to catch these
problems at compile time.