Fixed PairCorrEstimator so that it will pass unit test.
The reason it failed was that norm_factor was built using two approximations:
1.) Small delta approximation: (r+delta)^3 - r^3 = 3*delta*r^2 + 3*r*delta^2 ~ 3*delta*r^2
2.) Large N approximation: N(N-1) = N^2 - N ~ N^2

Both approximations are badly violated for a small cell with a large delta and small number of particles,
which is exactly the case for the unit test.

Because computing the normalization is not the computational bottleneck in a typical QMC calculation,
I changed norm_factor to use the exact expressions.