Update HDF5 not found message.
The suggested HDF5 version is out of date.  Version 1.8 or newer is needed,
or the code will not compile.  The HDF5_VERSION variable does not work
with 1.6 - checking the version is not going to be useful.  The last 1.6
version was released in 2009, so this version is unlikely to be an issue
anyway.