Removed the use of const h5_observable_type:
(1) the datatype should have been hid_t, not hsize_t

(2) In C++, for global objects, the constructor is called before main() is called. This results in H5open() getting called before MPI_init, which is not good practice in HDF5 usage. This also causes failure in the DAOS VOL connector call daos_cont_global2local. This failure is a result of the DAOS vol connector being initialized in serial and then later used in parallel.

Changed to pass H5T_NATIVE_DOUBLE to H5Dwrite instead.