Move stack sample storage to a separate class
The SampleStack class now holds the list of samples for later in optimization.

Some methods like loadEnsemble stayed in MCWalkerConfiguration because
they depend on details of MCWalkerConfiguration, but only on the
external interface to SampleStack.

The actual storage in the MCSample structure should be isolated so the
details can be changed (to SoA) if necessary.