Adjust RNG life cycle in QMCDriverNew
For optimization (and other cases where the driver is reused), returning
the RNG's to global control at the destructor is not the correct
lifecycle.

Add 'endSection' to QMCDriverInterface to allow drivers to release
resources, etc, that have been allocated or set up in earlier functions
(most likely in 'put' or 'process')
The QMCMain::runQMC routine contains the sequence of calls that create
the driver, set it up, run it, etc.

Increase the size of the RngCompatibility vector to the number of
threads in order to use it with old drivers (for the optimizer - to mix
the new driver to generate the samples with the old driver to evaluate
the samples)