Add QMCOptimize driver using batched VMC
The QMCOptimize driver was copied to QMCOptimizeBatched and converted to
work with batched VMC for the collection of samples.  This is intended to
have the minimal number of changes needed to compile and run without crashing.

This code only works for one thread.  The old VMC driver collected
samples on each thread.  The new VMC driver collects a single vector of
samples.  A future commit will distribute the samples to the per-thread
structure to enable threaded evaluation of the cost function.

There are a couple of TODO items in QMCOptimizeBatched::generateSamples
where the calls on vmc engine for the new batched driver are commented
out.  The effects of those calls should be investigated further to
determine if they need replacement functionality.