Fixes for gradient batched optimizer
- Change the input name from "opt_batched" to "opt_batch" to match "vmc_batch"

- Populate RngCompatibility for use in QMCCostFunction.  The RNG state
gets reset so the spherical integral grid for the non-local PP uses the
same grid during correlated sampling.  Even if the there is no non-local
PP, the RNG state movement in QMCostFunction needs the RNG's to exist.

- Create clones for use in QMCCostFunction.  The first element of
wClones and hClones is the input W and H value, but makeClones still
needs to be called for this to get set up (and for wClones/hClones to
have greater than zero length).

- Add method to set variational parameters in the per-walkers copies in MCPopulation

The test for this item is to compare 'Eavg' in the output text with the
results in the .scalar.dat files.  The former value comes from the evaluation
of the energy in QMCCostFunction::checkConfigurations, which uses the main Psi.
This currently is updated by the optimizer loop.

The latter value (.scalar.dat) comes from the VMCBatched run, which uses
the per-walker copies of the wavefunction