Add flex version of evaluateDerivatives
Make it clear that this function evaluates derivatives of the
variational parameters (vs. spatial derivatives) by adding 'Parameter'
to the name.

The original version works on one configuration and returns an array with
one element per variational parameter (well, two arrays).   The flex
version will evaluate several configurations, and needs to return an
array of arrays.  Possible solutions are to use a Matrix, use nested
std::vector, or create a new data structure.   We use the last approach,
the RecordArray class.  It should provide some abstraction between
the structure and the storage used.