Allow complex build to work with open BC
The unit test uses atomic orbitals, and fails with the complex build
in EvalPeriodicImagePhaseFactors with the "Attempting to run PBC LCAO
with no HDF5 support" error.
For actual runs, the complex build should not be used with LCAO and open
boundary conditions (??).  If we want to enforce this, it would be nice
to have an error message stating that.