Add check_scalars input options
Add --name, --ref-value, and --ref-error to check arbitrary named
columns in a scalar.dat file (versus adding the names to 'quantites'
list.

Implements #2635

This change only accepts a single value to test. If someone wants to test
multiple values in a single invocation, the 'action' value on add_option
could be set to 'append', and then the variables would become a list.