Fix accept ratio and block cpu time in batch vmc
Items in the estimators are divided by the block weight in
EstimatorManagerNew::makeBlockAverages.   The acceptance ratio
should not be divided by the block weight before putting into
the estimator manager.
The block CPU time needs to protected from that divide by multiplying
by the block weight before putting it into the estimator manager.