Add initial parameter gradient test
The system is a Helium atom with a bspline e-e jastrow with 4 parameters.

The test uses a new QMC_RUN_AND_CHECK_CUSTOM_SCALAR function.
This test uses the --name, --ref-value, --ref-error options to
check_scalars.py added in #2640 in order to test arbitrarily named
values.
The new function uses named arguments that are handled with
cmake_parse_arguments. This should make it easier to add new
functionality via new named arguments in the future.