Fix call to GradCost in complex build
In the complex build, the call to GradCost in testDerivOptimization was
calling the empty base class version QMCCostFunctionBase due to a
parameter mismatch.
Add 'override' to catch the issue, fix the signature, and add more
override's to keep clang happy.

Change a parameter to const and pass-by-reference in
in testDerivOptimization.

Add const to the print function in VariableSet (from debugging a
Jastrow optimization issue).