Set the RNG state during cost function evaluation
Set the RNG state to be the same during checkConfigurations and
correlatedSampling.

The "test" optimize method compares the analytic and finite difference
parameter gradients.  If the RNG is not reset before correlatedSampling,
the gradients can differ in the comparison by 1-10%.   If everything is
correct, and the RNG is reset properly, the gradients differ on the order
of 1e-8 to 1e-6.