Hack for correct VMC timer ordering
The VMC timers for these objects have the same scope as the object
lifetime (that is - timer starts in the constructor, and stops in the
destructor).
The call to make_unique creates the new object first, and then deletes
the old object.  This causes an incorrect ordering in the timers.
(The new timer starts before the old timer stops.)
This results in an error message - "Timer stack pop not matching push".
The printed timer results also show VMC nested under VMC.

The fix is to force the old object to destroyed before the new object is
created.