Update VMC energy for batch opt.
This difference in energy between batch and non-batch VMC only came up
after the linear batch optimizer was changed to use QMCDriverNew as its
base class.
The energies during optimization are the same.  It's only the VMC that
happens after the optimization loop that is different.

Probably something with the handling of the RNG in QMCDriverNew vs.
QMCDriver, but I didn't track down the root cause.