Skip updating parameters if outputing matrices
If the linear method has output matrices enabled, assume that we are
trying to accumulate points to compute statistics in the output file.
So each line in the scalar.dat file should use the same parameter
values.

In the future one might want to track the matrices as the optimization
proceeds.  This could then be moved to a separate option.