Use parameter input rather than attributes
To enable output of linear matrices, use the parameter tag rather than
an attribute.

Also add a separate parameter to freeze the change in variational
parameters.  It is clearer that this effect is desired.

For output of linear matrices for computing statistics and comparison,
the input file now looks like:
      <parameter name="output_matrices">yes</parameter>
      <parameter name="freeze_parameters">yes</parameter>