Fix indexing in diff Jastrow component
Second part for fix of #2814

The myVars in this object contains all the variational parameters
aggregated from all the radial functions (in their myVars), including
the non-active ones.

The OffSet variable maps the indices from the radial function myVars to
myVars in this object.

The output for parameter derivatives should use a contiguous index (i.e.
all the non-active parameters removed).  This is what the Index value in
VariableSet maps (returned by myVars.where).
In the code, index k covers all variables in myVars.  The index kk is
the contiguous index for output.

Some of the arrays are indexed with k, but should be indexed with kk
instead.