Fix two body Jastrow optimization
Second part of fix for #2814

Construct a second mapping for component-level variational parameters to
global-level variational parameters.

One potentially signficant change is that this maps the parameters as a block
from the component level to the global level.  The previous code could be more
fine-grained.
If there are cases where the subcomponent variables could be a mix of
active and inactive, then this code will need updating.