Fix 3172 - parameter derivatives in CI coeff
The temporary variables q0 and v need to be defined in the right scope
to get reset to zero.

Add unit testing that would catch the issue.

Change the signature of evaluateDerivatives in MultiSlaterDeterminant (not Fast)
to also call that routine in complex builds. (Otherwise the generic version in
WaveFunctionComponent gets called, and the derivatives are zero).