Fix batch optimizer issue with accepting moves
In QMCCostFunctionBatched::correlatedSampling, there is code to trim
extreme weights.  The weight calcution uses "inv_n_samples"
during accumulation.  That should use the global number of samples
(the number of samples across all MPI ranks), not the local number of
samples.

Add the global number of samples to the SampleStack class.

Adding the number of ranks to setMaxSamples is intended to keep the object
consistent between the number of samples and the global number of samples.
Inconsistent state is possible in the case of a separate function to
set the ranks or global number of samples.
However, setMaxSamples also allocates storage, and the number of ranks
has nothing to do with that.  From that perspective, setting the number
of ranks/global number of samples should be a separate function.