Possible fix for energy density
Addresses two issues causes for #2994

One is the loop bounds in QMCHamiltonian::initialize_traces
starts at 1 rather than 0. This would skip the kinetic energy,
an operator in the quantum domain would not be in the loop,
and a trace of Vq would not requested.

The second is ensuring the per-particle S(k) state is initialized for
the traces. Otherwise there is a segfault from eikr_r/i being
zero-length vectors.