Also handle CoulombPBCAB
Turn on the per-particle SK code when tracing.   It is unlikely this
code will ever get hit because the per-particle code would already have
been activated in the e-e CoulombPBCAA and the ion-ion CoulombPBCAA.
However, edge cases can be constructed (like leaving out the e-e or
ion-ion interaction) that would cause the code to segfault later.