Add member for number of node-local samples.
Add a member variable (numSamples_) to track the number of samples local
to each node.  Previously the number of samples was obtained from the
SampleStack variable (samples_) every time.  For testing this required
full initializing of the SampleStack.  Now, depending on the function
being tested, it may only requires initialization of the numSamples_ variable.

Sequence of calls:

  - constructor (SampleStack does not have any collected samples at this
point, don't know actual number of samples)

  - getConfigurations - initializes dLogPsi and d2LogPsi (used to get WF
derivatives). These are sized to the number of samples.

  - checkConfigurations - initializes RecordsOnNode, DerivRecords, and HDerivRecords
    and fills those in via WF and Hamiltonian evaluation (which use samples)

  - fillOverlapHamiltionMatrices compute the linear matrices from
      RecordsOnNode, DerivRecords, and HDerivRecords
      The samples are not used here. The number of samples is needed for
      iterating over RecordsOnNode, etc.