Add HDF file for inputs to linear method matrices
Add the parameter "output_matrix_inputs_hdf" to write out the per-sample
values for the wavefunction values, derivatives, Hamiltonian
derivatives, etc.   These values are used in
fillOverlapHamiltonianMatrices.  Along with the file from
"output_linear_matrices_hdf", these files can be used in the fillFromH5
unit test.
One set of files is included in this check-in. It should be
straightforward to create additional sets of files for testing locally.
(We don't want to increase the size of the repository unnecessarily)