update unit and converter tests to reflect changes in MSD for dirac
I updated the gold.orbs.h5 to use convert4qmc -dirac df_Bi.out -TargetState 14 -prefix gold -nojastrow
This tests the converter to make sure it can grab an arbitrary excited state to store in the hdf5.
However, this change made it so that I needed to update the multi_slater unit tests, which relied
on this gold.orbs.h5 file. Also, added to the unit test is reading the CI coeffs directly from the h5 file