Guard against null pointer use with nonlocalpp
In the case where the input file specifies nonlocalpp for
optimization but the system is not using pseudopotentials, nlpp will be
null, and the access will cause a segfault.
Ignoring the input value reproduces the behavior of the non-batched cost
function.

Also enable output earlier to print out messages about whether the non-local
piece is found or not.