Save and restore variational parameters to XML
The first step towards a solution to save and restore all variational
parameters (VP).

Reading from a VP file is not intended to create the list of
variational parameters.  The usage model is that the variational
parameters are set up as usual in the code, and then new values for
those parameters are read in from the file.

This enables a workflow where some subset of parameters is optimized
and then the wavefunction is enlarged (or the optimization of more
parameters enabled) and further optimization is performed. Values that
are not present in the VP file are not changed. Reading the VP file acts
like an overlay of its values on the existing values read in from the
wavefunction section.

The the list of values is contained under a "name_value_pairs" tag.  In the
future one might want to store values differently, and that could be
done under a different tag. For example, coefficients are often stored
as Coeff_0, Coeff_1, etc. for a large number of values. It would take less space
to store the prefix once and store all the values in an array.