Test loading of variational parameter files.
To avoid storing an HDF file in the repository, this test uses a script
that converts from a text format to the HDF format.
The convert_vp_format.py script can also be useful for editing the
variational parameter files (convert HDF to text, edit, and convert text
to HDF)

The complex build will store the variational parameters as complex
numbers, and expect to load them as complex numbers.
Because of this, the variational parameter files will not be compatible
across real and complex builds.
The script deals with this by adding a "--complex" flag that will output
the values in the complex number format.  The script can also be used
to convert between the real and complex versions of the variational
parameter files.  It will detect and read text or HDF files in either number
format.  The "--complex" flag only affects output.