Use walkers_per_crowd to match VMC
The size of walkers_per_crowd is the number of crowds and each element
is the batch size of each crowd.
The number of walkers per rank (sum of elements of walkers_per_crowd)
controls the memory usage on the device, which is the primary
constraint.

This matches how VMC works and divides walkers and crowds.

Previously the code expected the batch size of each crowd to be the
same.