Set intermediate indices in myVars
Ensure checkOutVariables gets called on the wavefunction clones.
This is needed to set the local to global index mapping on all the leaf
wavefunction components.
Need to also add it in getConfigurations so it gets set on the first round
of optimization.

The local myVars object is copied during cloning, but the clones get made
before myVars gets set (via checkOutVariables) in the primary wavefunction.