Build rotation variables only once.
If buildOptVariables gets called again, it will clear myVars.
The global indexing that gets set by checkOutVariables gets cleared as
part of this, and doesn't get set up again by buildOptVariables.
This will cause clones to write to the wrong location in dlogpsi and
dhpsioverpsi in evaluateDerivatives.

This code uses the size of myVars as a flag for rebuilding.
It might be possible to use the nel_major test (change to "nel >=
nel_major_") for the same effect. But I'm not sure what that variable is
supposed to be for, and it seems better to add a separate test.