Apply rotation and save history
The problem with previous code is that checkInVariables is used to
signal the beginning of an optimization cycle.  It does not get called
on all the clones, so it does not work well as this signal.

The rotation code depends on two things in the call to checkInVariables:
1. Sets the parameters to zero in order to get the parameter change when
resetParameters gets called
2. Stores the rotated coefficient matrix for later use.

The new code addresses these as:
1. Computes the delta between old parameters (in myVars) and the new
parameters (in active).  This also has the effect that calling
resetParameters multiple times with the same parameters will have no
effect.
2. Always applies the rotation from the existing coefficient matrix -
nothing is saved.   This should work for both the LCAO and spline code.

One problem that arises is the coefficient matrix is stored as a shared
pointer, and is shared among the clones.  If multiple threads call resetParameters,
the rotation gets applied multiple times.
The extreme hack to fix this here is to use the fact that checkInVariables is in
fact only called on one clone, and use that clone to peform the rotation.

This should be changed to something better before going into the code.

A history of the applied rotations is saved in resetParameters
and later code will save and restore it.