Improve gradient test output
Allow changing the gradient test finite difference delta from the input file.
Structure the output of the table so the column headers are aligned with the columns.
Increase the precision of the gradients and decrease the precision of
the relative difference.
Add thresholds and print !, !!, or !!! after the relative difference to make it
easier to visually scan for problematic values.