Extract run_qmc to separate module
Generalize the run_qmc function to an arbitrary number of electrons
and extract it to a separate module.
Future validation code will use different wavefunctions and we want to
reuse the run_qmc function.

The implementation of psi in the wavefunction needs to be split into two
parts.  For generality, calling from run_qmc uses an array for all the
electron positions.  Autograd is easier to use if each position is a
separate argument.  The solution is to make psi a wrapper that extracts
individual positions and calls the internal implementation.