Fix deterministic tests.
The order of function argument evaluations is undefined in the standard.

This leads to a difference in evaluation order between debug and
release builds, which causes the order of rng1, rng2, and rng3 to be swapped.
In debug builds, the arguments get placed on the stack in the reverse
order listed.  In release builds, the code is likely inlined, and the
arguments effectively get evaluated in the forward order listed.