Protect the case when output size is smaller.
Fix SplineC2R. When the requested size is smaller than what can be supplied, only output the requested size.
What is the right way of selecting a subset of orbitals in an already built SplineC2R remains a big question.
The trucation is subject to the sorting of orbitals in C2R which puts all the "complex" orbitals first.
"complex" orbitals produce two real orbitals.