Make the coefficient copy a shared pointer
The copy of the cofficient matrix is used for orbital rotation.
The storage in this copy is used regardless of the flag setting.
For the global rotation method, the storage and the values of the copy
are used.  For the history list method, the storage is used, but
the values are overwritten.