Fixes to read VP.h5 files
Pass the VariableSet that was read in the main part of the vp.h5 file.
Need to set myVars in RotatedSPOs to its initial state w/o invoking
all the code in resetParametersExclusive.
Then subsequent calls to resetParametersExclusive will compute the
parameter delta correctly.

Also add a check in the derivative computation that the index is valid.
The index might be invalid (-1) if the rotational parameters are not
included in the variational_subset being optimized.