Use name of rotated SPOset in h5
When naming the rotation datasets in the VP h5 file, use the rotated
SPOSet name rather than the underlying SPOSet name.
The underlying SPOSet name isn't checked for uniqueness, and a mistake
in the input file could result in duplicate SPOSet names, which then
gives only a single entry in the VP h5 file, when there should be two.