Read and write rotation parameters
Add functions for applying either global rotation or history list.
The choice of history list vs. global rotation for a given run
can be independent of how is was stored in the VP file.
That is, the method used to restore the parameters in RotatedSPOs
is determined by what method is stored in the VP file.
Once the parameters in the RotatedSPO class are set up, it does not matter
what method future optimization runs use.