Perform group check inside hdf_archive
Addresses issue #4547 (handle h5 group check in hdf_archive)

When 'push' is called with the createit flag set to false, and the group
does not exist, throw an exception.

Also throw an exception if the file is not open.

Remove the return value (use 'top' to get the value if it's really
needed)

Store the hdf archive filename so it can be included in the error message
in the exception.