Updates to filename and group handling
Clear the current filename in the close() call.  That also means moving
the statement to set the filename after the close() calls in open and
create.

Add a function to get a string representation of the current group
stack, in order to enhance the error message when a group is not found.
A vector is used to store the group names instead of a stack because
it's difficult to iterate over a stack.