Fix assertion failure.
Addresses #4666.
mw_evaluate_notranspose and evaluate_notranspose don't produce bitwise identical results.
The problem started by https://github.com/QMCPACK/qmcpack/pull/4407 which introduce a batched evaluation and caused bitwise matrix value checking failure. I printed diff and found the value is at the epsilon of float or double. So no concern.

The intention of the assertion was to catch any missing D2H transfer earlier and there is no need of recomputing. So the solution is removing the re-computation.