Workaround for clang with glibc 12 include files
Deals with clang as the compiler in CUDA mode with gcc/glibc 12 include files.
Clang was fixed to recognize __noinline__ as an attribute,
but the CUDA include files still have a define for __noinline__.
The attribute started to be used in gcc 12 include files, which expands
to __attribute__((__attribute__((noinline)))), which the compiler rejects.
See https://github.com/NVIDIA/thrust/issues/1703 and
    https://github.com/llvm/llvm-project/issues/57544