Additional zero/NaN protection in MSD.
When the proposed configuration is on the nodal surface.
Reference determinant curRatio goes to zero.
This move should be rejected eventually by Monte Carlo.
However, to make the code proceed, some NaN protection is needed.
When Dets[det_id]->getRefDetRatio() is zero, new_psi_ratio_to_new_ref_det_ the sum of all the dets
with respect to the ref one is NaN because it is the sum of all the 1/0.
So keeping curRatio zero should be healthy.