Add adaptive rc
Sometimes a crossing in the 2nd derivative is not found.  In that case,
increase the initial fitting range and try again.

n may not be quantum number, but we can treat it like one, especially
for 1s and 2s orbitals.

Move the lower bound for computing the reference values to 0.1.
That is, the range of values used for the initial fit is [0.1, 0.2].
This seems to work better.