Fix occasional optimizer failure.
Sometimes no eigenvalue is found from the mapping (filter and sort), which
leads the code to select the first eigenvalue returned from the solver,
which often sends the optimizer parameters in a very undesirable direction.

The workaround is to detect the situation where no acceptable eigenvalue
is found, and run a second filter which accepts eigenvalues above the
current VMC energy ( H(0,0) ).
The solution could be simpler, but the fix was chosen to keep the
existing eigenvalue selection unchanged as much as possible.