consistent init and use input name and type
What I believe the intention of the stat.h5 output is that if a name
is supplied for an estimator than that will be its group name in the
h5 file.  If no name is supplied the estimator type will be
copied to the name and that will be used.