Remove Deprecated NumPy Aliases
The NumPy aliases `np.string_` and `np.float_` were deprecated in NumPy 2.0.0. The correct aliases have been available since NumPy 1.13, and therefore are unnecessary. Further, the function for numpy compatibility in `testing.py` was removed and the command for altering print settings was placed in the try-except statement. All instances where `np.bytes_` were added to strings were altered so the `np.bytes_` object was first converted to string to ensure compatibility.