[Nexus] Remove unused keyword argument in `plotting.py`
After recent addition of explicit error handling an error in `plotting.py` was shown to arise from attempting to pass a keyword variable `warn=False` in line 21 of `plotting.py`. This raised a `TypeError` that previously went uncaught as the `try... except` statement lacked an explicit set of errors to catch. Removal of this unused keyword argument should fix the issue.