[Nexus] Remove deprecated keyword `warn` from `plotting.py`
The module `plotting.py` contained a series of `try... except` statements to test for matplotlib, however in the innermost nested `try... except` the command `matplotlib.use()` was invoked with a keyword `warn=False` that has been deprecated since at least the release of matplotlib 3.3, and has been defaulted to `False` since matplotlib 3.0. Given that matplotlib <3.0 is for Python 2 mostly, and since Nexus does not support Python 2 anymore, this should be removed.

Previously, this was throwing an exception that was not caught because of the `try... except` statement it was contained in that caught all exceptions, in this case a `TypeError`, even though it should only be catching the correct exception of `ValueError`.