[Nexus] Import Restructure and Cleanup
This is the first pass at Issue #5620, designed to remove unused imports, clean up namespace, and generally improve readability of functions from external packages. This commit changes only the file `structure.py` and, barring significant error in testing, should serve as an example for the rest of the import restructuring efforts.