[Nexus] Correct `Structure.center_molecule()` method
Fixed an error regarding the use of the cell center to shift the origin of the system. Corrected to use the difference (if any) between the cell center supplied by the user and the calculated cell center. If the user does not supply a center this is equivalent to the point (0, 0, 0) and thus will not shift the molecule at all.