[Nexus] Correct modified `parse_string()` function in `xmlreader.py`
The `parse_string()` function that was added in `xmlreader.py` did not match previous test expectations. The function would make integers and integer arrays more frequently than the previous method since it checked for values matching the mathematical definition of integer (i.e. 1.0 is an integer mathematically but is a float in computer science). This would create issues as calling `np.array(data, int)` on things that were not necessarily all written as integers in a string would cause a failure, similar to Issue #5626. The code now has been updated to conform to the expectations set by `test_string2var()` that previously existed in `test_superstring.py`.