[Nexus] Fix bug in `qmcpack_input.py`
A bug arose in the fix of `qmcpack_input.py` that was due to the handling of errors in lines 1581-1584. The reason for this bug was that the error from the attempts to convert text to an array was always caught, and at the end a `pass` was given that caught the last error, so the first part of the larger `try... except:` block would not see that an error occurred previously. This fix removes the inner nested `try... except:` that caught the error, and now the error should be caught by the outer `try... except:` and handled appropriately.