[Nexus] Fix incompatibility with NumPy converters in `numpy.loadtxt()`
An interesting quirk of `numpy.loadtxt()` was discovered that previously was not noticed.

The function would load text files and if the parameter `dtype=int` was passed, would load the file using the parsing ability of `float()` and then cast the result as `int()`, meaning that information would be lost if the actual data type was `float` and not `int`.

The reason this was not spotted before is that somehow the `is_int()` function that was previously implemented to check for integer arrays would seemingly never actually detect an integer and try to load the text as an `int`, or alternatively would fail and get handled somewhere else.

The current fix still implements the changes in the rest of the PR, however it now specifies that the converter used should be `int` and not `float`. This should be backwards compatible with all supported versions of NumPy.