[Nexus] Requested changes for PR #5627
This commit makes the changes requested by @jtkrogel regarding function definitions and preserving testing for the functions that previously were in `superstring.py`.

`valid_variable_name()` was added to `developer.py` and the corresponding test was added to `test_developer.py` and included in `nxs-test`.

The functions `parse_string()`, `find_pair()`, `remove_pair_sections()`, and `remove_empty_lines()` were defined if not already in `xmlreader.py` and added to `test_xmlreader.py` as well as included in `nxs-test`. In addition, the function `find_pair()` and `remove_pair_sections()` were updated to indicate that they accept `tuple[str, str]` instead of `list[str, str]`, and the existing places in `xmlreader.py` where `list[str, str]` was passed have been updated to pass `tuple[str, str]`.

The change regarding `split_delims()` in `pseudopotential.py` was changed again to use the already existing and tested `pp_elem_label()` in `pseudopotential.py`.

The replacement of `valid_variable_name()` in `check_stats.py` was updated so that the function is now in the `developer.py` block.